Add SDR for 978MHz (#36)
* Add SDR for 978MHz

I tried to make this shorter than the first version, but it's still too much. Feel free to use as much or as little of this as you wish.

* grammar

"GA aircraft are reporting" should read "GA aircraft that are reporting"

* fixed rejected typos

Automated merge testing failed for typos. Cannot fix METARs.

* revert typo fix

The automatic spelling checks keep rejecting "multilateration" but that's what MLAT is.