Small suggestions (#75)
* Add GitHub links to How to Get Help section

While this may be redundant to links elsewhere, I always feel that the easier
it make for people to interact, the likelier they are to contribute.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>

* Add another argument for using Docker

Again, not sure I can claim this is needed - but it's one of those significant
game changers that people often overlook. All you do is edit this one file.
At least to me, that's brilliant.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>

* Small edits to the equipment

It seems useful to point out that outdoor antennas tend to give more joy.
It seems useful to acknowledge that Pis are currently hard to get. Maybe even a link
to rpilocator.com.
I certainly wasn't aware that this existed before getting into this hobby.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>

* Reorganized the information needed

Since the elevation finder already gives you the lat/long in sufficient
precision, it seems to make more sense to start with that and avoid
having people first hunt for their coordinates and then for their
altitude, only to get their coordinates from that as well...

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>

* Add new words to spell checker dictionary

* Fix some spelling

* Update .spelling

* Spelling

* Linting

* Linting

* Linting

* Linting

---------

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>
Co-authored-by: Mike <mike.nye@gmail.com>