Add suggestions for how to deal with not seeing any planes
In very RF noisy environments, the starting value of max gain for the
autogain script can cause no ADS-B messages being received at all, which
looks like the project not working to the user. And because of the
implementation of the autogain script, prevents it from trying lower
gain values.
I wasn't clear how much detail we want to give here - but this should
get people started.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>